requirements.txt
decorator==4.0.6
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5
python-dateutil==2.4.2
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wheel==0.26.0
